+++ scrapy_cffi/utils/__init__.py
@@ -48,7 +48,11 @@ if TYPE_CHECKING:
         safe_call,
     )
     from .email import Email
-    from .envConfig import env_to_settings, settings_to_env
+    from .envConfig import (
+        env_to_settings,
+        load_env_settings,
+        settings_to_env,
+    )
     from .fd import FDUtil
     from .jsonLoad import JSONExtractor, JSONScanner, extract_json_chain, extract_nested_objects
     from .log import (